Create Pixmap color blend tests
You'll need to change the package import and class names to whatever project you plug this into as well as define the dimensions of the window in the main class to be 500x500, but other than that this code should be the stand alone class you need to run what I was testing. It should display a 500x500 pixel screen with each of the four corners being completely one color and gradient fading as the color gets further from its origin corner. The top-left should be red, top-right should be blue, bottom-right should be green, and bottom-left should be alpha.